Financial Analysis API (Backend)/requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai
yfinance
pydantic
apscheduler
beautifulsoup4
